Challenge1App/Challenge1App/Employee.cs
namespace Challenge1App
{
    public class Employee
    {
        private List<int> score = new List<int>();
        public Employee(string name, string surname, string age)
        {
            this.Name = name;
            this.Surname = surname;
            this.Age = age;
        }
        public string Name { get; set; }
        public string Surname { get; set; }
        public string Age { get; set; }
        public int Result
        {
            get
            {
                return this.score.Sum();
            }
        }
        public void AddScore(int number)
        {
            this.score.Add(number);
        }
        public void DeleteScore(int number)
        {
            this.score.Remove(number);
        }
    }
}
